Split the 'internal' module to unix and windows sub-modules.
