+++ src/internal/mod.rs
@@ -11,15 +11,18 @@ extern crate libc;
 use std::io;
 use std::mem;
 
-pub use self::native::{addr_to_sockaddr, close, retry, sockaddr_to_addr};
+pub use self::native::{addr_to_sockaddr, sockaddr_to_addr};
+
+#[cfg(windows)]mod windows;
+#[cfg(not(windows))]mod posix;
+
+#[cfg(windows)]
+pub use self::windows::*;
+#[cfg(not(windows))]
+pub use self::posix::*;
 
 mod native;
 
-#[cfg(windows)] pub type CSocket = libc::SOCKET;
-#[cfg(windows)] pub type BufLen = i32;
-
-#[cfg(not(windows))] pub type CSocket = libc::c_int;
-#[cfg(not(windows))] pub type BufLen = libc::size_t;
 
 // Any file descriptor on unix, only sockets on Windows.
 pub struct FileDesc {
